refactor: move categories into blueprint set
