+++ views/main.py
@@ -1,3 +1,8 @@
+"""
+Main View Module
+"""
+
+
 from flask import redirect
 from . import main_bp
